CPAN version 0.03
Determine whether we have an internet connection before running tests.

Internally we try to open an internet connection as part of constructing
a Perl::RT2Github.pm object.  In most circumstances this is perfectly
reasonable, but some CPANtesters rigs appear to test without internet
connections and so are failing t/new.t.  With this version, we attempt
to work around them.

No change in functionality.

Update Changes and README.